+++ CHANGELOG.md
@@ -2,13 +2,17 @@
 
 ### Improvements
 
+## 0.22.0 (Released January 27, 2021)
+
+### Improvements
+
 - Initial support for Go
   [#519](https://github.com/pulumi/pulumi-eks/pull/519)
 
 - Initial support for Python and .NET
   [#448](https://github.com/pulumi/pulumi-eks/pull/448)
-  
-- Add support for `kubernetesServiceIpAddressRange` to `eks.Cluster` 
+
+- Add support for `kubernetesServiceIpAddressRange` to `eks.Cluster`
   [#509](https://github.com/pulumi/pulumi-eks/pull/509)
 
 ## 0.21.0 (Released January 12, 2021)